Joaquin Alvarez - Updated the Menu System
I added in two control widgets for the control pages as well as added in a credit widget. I also added some functionality to most of the buttons. The music and sound effect slider are the only buttons that do not work right now. However, the might be buggy so I need to fix that.