Joaquin Alvarez - Updated UI
I started to add in the Boss UI. Not fully implanted yet, will add it and make it functional once the boss is added. I also fix the quit button location so it shows up on the title screen, I fix the size of the buttons, locations of buttons, and sound of the buttons for the menus.